Move implementation to cpp files
